requirements.txt
Django==2.0.7
psycopg2-binary
gunicorn
django-environ
djangorestframework
django-filter==2.0.0
django-cors-headers
django-debug-toolbar
antlr4-python3-runtime